Gherkin tests: Put features and functions related to groups in a different file. To slim the big app_language_groups.go file a bit more.
